requirements.txt
# PageIndex dependencies
openai>=1.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
Pillow>=9.0.0
python-dotenv>=1.0.0

# Document processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdown>=3.4.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6

# Testing
pytest>=7.0.0
